[SCRUM-150] Move : 이미지 관련 패키지 경로 변경
