+++ src/main/java/com/kkinikong/be/review/service/ReviewService.java
@@ -16,8 +16,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import com.kkinikong.be.global.response.PageResponse;
-import com.kkinikong.be.image.service.ImageService;
-import com.kkinikong.be.image.type.S3Bucket;
 import com.kkinikong.be.review.domain.Review;
 import com.kkinikong.be.review.domain.ReviewImage;
 import com.kkinikong.be.review.domain.ReviewTag;
@@ -41,6 +39,8 @@ import com.kkinikong.be.user.domain.User;
 import com.kkinikong.be.user.exception.UserException;
 import com.kkinikong.be.user.exception.errorcode.UserErrorCode;
 import com.kkinikong.be.user.repository.UserRepository;
+import com.kkinikong.be.util.s3.service.ImageService;
+import com.kkinikong.be.util.s3.type.S3Bucket;
 
 @Service
 @RequiredArgsConstructor
